requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg[binary]==3.2.3
python-dotenv==1.0.1
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
langchain==0.2.1
langchain-community==0.2.1
langchain-huggingface==0.0.3
huggingface-hub==0.23.2
pypdf==4.2.0
python-docx==1.1.2
pillow==10.3.0
pytesseract==0.3.10
httpx==0.27.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1